security-package/issues/86: ReCaptcha configuration refactoring
