+++ ReCaptchaAdminUi/composer.json
@@ -6,8 +6,7 @@
         "php": "~7.1.3||~7.2.0||~7.3.0",
         "magento/framework": "102.0.*",
         "magento/module-config": "*",
-        "magento/module-re-captcha-api": "*",
-        "magento/module-re-captcha-ui": "*"
+        "magento/module-re-captcha-api": "*"
     },
     "authors": [
         {
