+++ ReCaptchaFrontendUi/composer.json
@@ -7,8 +7,7 @@
         "magento/framework": "102.0.*",
         "magento/module-store": "*",
         "magento/module-config": "*",
-        "magento/module-re-captcha-api": "*",
-        "google/recaptcha": "^1.2"
+        "magento/module-re-captcha-api": "*"
     },
     "authors": [
         {
