+++ ReCaptchaNewsletter/Observer/NewsletterObserver.php
@@ -13,7 +13,7 @@ use Magento\Framework\Event\Observer;
 use Magento\Framework\Event\ObserverInterface;
 use Magento\Framework\Exception\LocalizedException;
 use Magento\ReCaptchaApi\Api\CaptchaConfigInterface;
-use Magento\ReCaptchaUi\Api\RequestHandlerInterface;
+use Magento\ReCaptchaUi\Model\RequestHandlerInterface;
 
 /**
  * NewsletterObserver
