+++ ReCaptchaUi/Model/RequestHandler.php
@@ -17,7 +17,6 @@ use Magento\ReCaptchaApi\Api\CaptchaConfigInterface;
 use Magento\ReCaptchaApi\Api\CaptchaValidatorInterface;
 use Magento\ReCaptchaApi\Api\Data\ValidationConfigInterface;
 use Magento\ReCaptchaApi\Api\Data\ValidationConfigInterfaceFactory;
-use Magento\ReCaptchaUi\Api\RequestHandlerInterface;
 
 /**
  * @inheritdoc
