+++ ReCaptchaUi/etc/di.xml
@@ -7,5 +7,5 @@
  -->
 <config xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
         xsi:noNamespaceSchemaLocation="urn:magento:framework:ObjectManager/etc/config.xsd">
-    <preference for="Magento\ReCaptchaUi\Api\RequestHandlerInterface" type="Magento\ReCaptchaUi\Model\RequestHandler"/>
+    <preference for="Magento\ReCaptchaUi\Model\RequestHandlerInterface" type="Magento\ReCaptchaUi\Model\RequestHandler"/>
 </config>
